improved GUI
